Implemented Keyring subclass
