+++ test/packages/caver.transaction/feeDelegatedAccountUpdate.js
@@ -33,6 +33,7 @@ const { propertiesForUnnecessary } = require('../utils')
 const testRPCURL = require('../../testrpc')
 const Caver = require('../../../index.js')
 const Keyring = require('../../../packages/caver-wallet/src/keyring/keyring')
+const SingleKeyring = require('../../../packages/caver-wallet/src/keyring/singleKeyring')
 const TransactionHasher = require('../../../packages/caver-transaction/src/transactionHasher/transactionHasher')
 
 const { generateRoleBasedKeyring, makeAccount, accountKeyTestCases, checkSignature, checkFeePayerSignature } = require('../utils')
@@ -527,11 +528,11 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-162: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
 
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
-                await tx.signWithKey(sender.keys[0][0].privateKey)
+                await tx.signWithKey(sender.key.privateKey)
 
                 checkFunctionCall(i)
                 checkSignature(tx)
@@ -541,7 +542,7 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-163: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
 
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
@@ -681,10 +682,10 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-171: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
-                await tx.signFeePayerWithKey(feePayer.keys[0][0].privateKey)
+                await tx.signFeePayerWithKey(feePayer.key.privateKey)
 
                 checkFunctionCall(i)
                 checkFeePayerSignature(tx)
@@ -694,7 +695,7 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-172: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
                 await tx.signFeePayerWithKey(feePayer.getKlaytnWalletKey())
@@ -818,10 +819,10 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-179: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
-                await tx.signWithKeys(sender.keys[0][0].privateKey)
+                await tx.signWithKeys(sender.key.privateKey)
 
                 checkFunctionCall(i)
                 checkSignature(tx)
@@ -831,7 +832,7 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-180: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
                 await tx.signWithKeys(sender.getKlaytnWalletKey())
@@ -944,10 +945,10 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-186: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
-                await tx.signFeePayerWithKeys(feePayer.keys[0][0].privateKey)
+                await tx.signFeePayerWithKeys(feePayer.key.privateKey)
 
                 checkFunctionCall(i)
                 checkFeePayerSignature(tx)
@@ -957,7 +958,7 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-187: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             for (let i = 0; i < txsByAccountKeys.length; i++) {
                 const tx = txsByAccountKeys[i]
                 await tx.signFeePayerWithKeys(feePayer.getKlaytnWalletKey())
@@ -982,7 +983,7 @@ describe('TxTypeFeeDelegatedAccountUpdate', () => {
                 await tx.signFeePayerWithKeys(roleBasedKeyring, customHasher)
 
                 checkFunctionCall(i, true)
-                checkFeePayerSignature(tx, { expectedLength: roleBasedKeyring.roleFeePayerKey.length })
+                checkFeePayerSignature(tx, { expectedLength: roleBasedKeyring.keys[caver.wallet.keyring.role.roleFeePayerKey].length })
                 expect(createFromPrivateKeySpy).not.to.have.been.calledOnce
                 expect(roleBasedSignWithKeysSpy).to.have.been.calledWith(hashForCustomHasher, '0x7e3', 2)
             }
