+++ test/packages/caver.transaction/feeDelegatedCancel.js
@@ -33,6 +33,7 @@ const { propertiesForUnnecessary } = require('../utils')
 const testRPCURL = require('../../testrpc')
 const Caver = require('../../../index.js')
 const Keyring = require('../../../packages/caver-wallet/src/keyring/keyring')
+const SingleKeyring = require('../../../packages/caver-wallet/src/keyring/singleKeyring')
 const TransactionHasher = require('../../../packages/caver-transaction/src/transactionHasher/transactionHasher')
 
 const { generateRoleBasedKeyring, checkSignature, checkFeePayerSignature } = require('../utils')
@@ -253,8 +254,8 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-382: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -263,7 +264,7 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-383: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -379,8 +380,8 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-391: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signFeePayerWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signFeePayerWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkFeePayerSignature(tx)
@@ -389,7 +390,7 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-392: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signFeePayerWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -492,8 +493,8 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-399: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -502,7 +503,7 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-400: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -596,8 +597,8 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-406: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signFeePayerWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signFeePayerWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkFeePayerSignature(tx)
@@ -606,7 +607,7 @@ describe('TxTypeFeeDelegatedCancel', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-407: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signFeePayerWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
