+++ test/packages/caver.transaction/feeDelegatedChainDataAnchoring.js
@@ -33,6 +33,7 @@ const { propertiesForUnnecessary } = require('../utils')
 const testRPCURL = require('../../testrpc')
 const Caver = require('../../../index.js')
 const Keyring = require('../../../packages/caver-wallet/src/keyring/keyring')
+const SingleKeyring = require('../../../packages/caver-wallet/src/keyring/singleKeyring')
 const TransactionHasher = require('../../../packages/caver-transaction/src/transactionHasher/transactionHasher')
 
 const { generateRoleBasedKeyring, checkSignature, checkFeePayerSignature } = require('../utils')
@@ -264,8 +265,8 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-454: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -274,7 +275,7 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-455: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -390,8 +391,8 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-463: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signFeePayerWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signFeePayerWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkFeePayerSignature(tx)
@@ -400,7 +401,7 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-464: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signFeePayerWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -503,8 +504,8 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-471: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -513,7 +514,7 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-472: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -607,8 +608,8 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-478: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signFeePayerWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signFeePayerWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkFeePayerSignature(tx)
@@ -617,7 +618,7 @@ describe('TxTypeFeeDelegatedChainDataAnchoring', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-479: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signFeePayerWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
