+++ test/packages/caver.transaction/feeDelegatedSmartContractDeploy.js
@@ -33,6 +33,7 @@ const { propertiesForUnnecessary } = require('../utils')
 const testRPCURL = require('../../testrpc')
 const Caver = require('../../../index.js')
 const Keyring = require('../../../packages/caver-wallet/src/keyring/keyring')
+const SingleKeyring = require('../../../packages/caver-wallet/src/keyring/singleKeyring')
 const TransactionHasher = require('../../../packages/caver-transaction/src/transactionHasher/transactionHasher')
 
 const { generateRoleBasedKeyring, checkSignature, checkFeePayerSignature } = require('../utils')
@@ -280,8 +281,8 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-237: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -290,7 +291,7 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-238: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -406,8 +407,8 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-246: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signFeePayerWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signFeePayerWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkFeePayerSignature(tx)
@@ -416,7 +417,7 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-247: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signFeePayerWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -519,8 +520,8 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-254: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -529,7 +530,7 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-255: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -623,8 +624,8 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-261: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signFeePayerWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signFeePayerWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkFeePayerSignature(tx)
@@ -633,7 +634,7 @@ describe('TxTypeFeeDelegatedSmartContractDeploy', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTIONFD-262: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signFeePayerWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
