+++ test/packages/caver.transaction/valueTransferMemo.js
@@ -33,6 +33,7 @@ const { propertiesForUnnecessary } = require('../utils')
 const testRPCURL = require('../../testrpc')
 const Caver = require('../../../index.js')
 const Keyring = require('../../../packages/caver-wallet/src/keyring/keyring')
+const SingleKeyring = require('../../../packages/caver-wallet/src/keyring/singleKeyring')
 const TransactionHasher = require('../../../packages/caver-transaction/src/transactionHasher/transactionHasher')
 
 const { generateRoleBasedKeyring, checkSignature } = require('../utils')
@@ -260,8 +261,8 @@ describe('TxTypeValueTransferMemo', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTION-115: input: private key string. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
-            await tx.signWithKey(sender.keys[0][0].privateKey)
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
+            await tx.signWithKey(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -270,7 +271,7 @@ describe('TxTypeValueTransferMemo', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTION-116: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeyProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKey')
+            const signWithKeyProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKey')
             await tx.signWithKey(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
@@ -374,8 +375,8 @@ describe('TxTypeValueTransferMemo', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTION-123: input: private key string. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
-            await tx.signWithKeys(sender.keys[0][0].privateKey)
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
+            await tx.signWithKeys(sender.key.privateKey)
 
             checkFunctionCall()
             checkSignature(tx)
@@ -384,7 +385,7 @@ describe('TxTypeValueTransferMemo', () => {
         }).timeout(200000)
 
         it('CAVERJS-UNIT-TRANSACTION-124: input: KlaytnWalletKey. should sign transaction.', async () => {
-            const signWithKeysProtoSpy = sandbox.spy(Keyring.prototype, 'signWithKeys')
+            const signWithKeysProtoSpy = sandbox.spy(SingleKeyring.prototype, 'signWithKeys')
             await tx.signWithKeys(sender.getKlaytnWalletKey())
 
             checkFunctionCall()
