+++ test/packages/caver.wallet.js
@@ -30,7 +30,9 @@ const expect = chai.expect
 const testRPCURL = require('../testrpc')
 
 const Caver = require('../../index.js')
-const Keyring = require('../../packages/caver-wallet/src/keyring/keyring')
+const AbstractKeyring = require('../../packages/caver-wallet/src/keyring/abstractKeyring')
+const SingleKeyring = require('../../packages/caver-wallet/src/keyring/singleKeyring')
+const MultipleKeyring = require('../../packages/caver-wallet/src/keyring/multipleKeyring')
 const PrivateKey = require('../../packages/caver-wallet/src/keyring/privateKey')
 
 const ValueTransfer = require('../../packages/caver-transaction/src/transactionTypes/valueTransfer/valueTransfer')
@@ -46,8 +48,14 @@ beforeEach(() => {
 })
 
 function validateKeyringInWallet(data, { expectedAddress, expectedKey } = {}) {
-    expect(data instanceof Keyring).to.be.true
-    const objectKeys = ['_address', '_keys']
+    expect(data instanceof AbstractKeyring).to.be.true
+    const objectKeys = ['_address']
+
+    if (data instanceof SingleKeyring) {
+        objectKeys.push('_key')
+    } else {
+        objectKeys.push('_keys')
+    }
 
     expect(Object.getOwnPropertyNames(data)).to.deep.equal(objectKeys)
 
@@ -58,15 +66,25 @@ function validateKeyringInWallet(data, { expectedAddress, expectedKey } = {}) {
     }
 
     if (expectedKey !== undefined) {
-        if (_.isArray(expectedKey) && _.isString(expectedKey[0])) expectedKey = [expectedKey, [], []]
-        if (_.isString(expectedKey)) expectedKey = [[expectedKey], [], []]
-        for (let i = 0; i < data.keys.length; i++) {
-            for (let j = 0; j < data.keys[i].length; j++) {
-                const privateKeyString = expectedKey[i][j] instanceof PrivateKey ? expectedKey[i][j].privateKey : expectedKey[i][j]
-                expect(data.keys[i][j].privateKey.toLowerCase()).to.equal(privateKeyString.toLowerCase())
+        if (data instanceof SingleKeyring) {
+            comparePrivateKey(expectedKey, data.key)
+        } else if (data instanceof MultipleKeyring) {
+            for (let i = 0; i < data.keys.length; i++) {
+                comparePrivateKey(expectedKey[i], data.keys[i])
+            }
+        } else {
+            for (let i = 0; i < data.keys.length; i++) {
+                for (let j = 0; j < data.keys[i].length; j++) {
+                    comparePrivateKey(expectedKey[i][j], data.keys[i][j])
+                }
             }
         }
     }
+}
+
+function comparePrivateKey(expected, actual) {
+    const privateKeyString = expected instanceof PrivateKey ? expected.privateKey : expected
+    expect(actual.privateKey.toLowerCase()).to.equal(privateKeyString.toLowerCase())
 }
 
 describe('wallet.generate', () => {
@@ -109,7 +127,7 @@ describe('wallet.newKeyring', () => {
         it('should create keyring instances with parameters and add to in-memory wallet', () => {
             const addSpy = sinon.spy(caver.wallet, 'add')
             const keyring = caver.wallet.keyring.generate()
-            const added = caver.wallet.newKeyring(keyring.address, keyring.keys[0][0].privateKey)
+            const added = caver.wallet.newKeyring(keyring.address, keyring.key.privateKey)
 
             validateKeyringInWallet(added, { expectedAddress: keyring.address, expectedKey: keyring.keys })
             expect(addSpy).to.have.been.calledOnce
@@ -161,40 +179,34 @@ describe('wallet.updateKeyring', () => {
     context('CAVERJS-UNIT-KEYRINGCONTAINER-007: input: coupled keyring', () => {
         it('should update key of keyring', () => {
             const coupled = caver.wallet.keyring.generate()
-            const copySpy = sinon.spy(coupled, 'copy')
             const decoupled = caver.wallet.keyring.createWithSingleKey(coupled.address, caver.wallet.keyring.generateSingleKey())
             caver.wallet.add(decoupled)
 
             const updated = caver.wallet.updateKeyring(coupled)
             const keyringFromContainer = caver.wallet.getKeyring(coupled.address)
 
-            validateKeyringInWallet(updated, { expectedAddress: decoupled.address, expectedKey: coupled.keys })
-            validateKeyringInWallet(keyringFromContainer, { expectedAddress: coupled.address, expectedKey: coupled.keys })
-
-            expect(copySpy).to.have.been.calledOnce
+            validateKeyringInWallet(updated, { expectedAddress: decoupled.address, expectedKey: coupled.key })
+            validateKeyringInWallet(keyringFromContainer, { expectedAddress: coupled.address, expectedKey: coupled.key })
         })
     })
 
     context('CAVERJS-UNIT-KEYRINGCONTAINER-008: input: decoupled keyring', () => {
         it('should update key of keyring', () => {
             const coupled = caver.wallet.keyring.generate()
-            const copySpy = sinon.spy(coupled, 'copy')
             const decoupled = caver.wallet.keyring.createWithSingleKey(coupled.address, caver.wallet.keyring.generateSingleKey())
             caver.wallet.add(coupled)
 
             const updated = caver.wallet.updateKeyring(decoupled)
             const keyringFromContainer = caver.wallet.getKeyring(coupled.address)
 
-            validateKeyringInWallet(updated, { expectedAddress: coupled.address, expectedKey: decoupled.keys })
-            validateKeyringInWallet(keyringFromContainer, { expectedAddress: coupled.address, expectedKey: decoupled.keys })
-            expect(copySpy).to.have.been.calledOnce
+            validateKeyringInWallet(updated, { expectedAddress: coupled.address, expectedKey: decoupled.key })
+            validateKeyringInWallet(keyringFromContainer, { expectedAddress: coupled.address, expectedKey: decoupled.key })
         })
     })
 
     context('CAVERJS-UNIT-KEYRINGCONTAINER-009: input: multiSig keyring', () => {
         it('should update key of keyring', () => {
             const coupled = caver.wallet.keyring.generate()
-            const copySpy = sinon.spy(coupled, 'copy')
             const multiSig = generateMultiSigKeyring()
             multiSig.address = coupled.address
             caver.wallet.add(coupled)
@@ -204,7 +216,6 @@ describe('wallet.updateKeyring', () => {
 
             validateKeyringInWallet(updated, { expectedAddress: coupled.address, expectedKey: multiSig.keys })
             validateKeyringInWallet(keyringFromContainer, { expectedAddress: coupled.address, expectedKey: multiSig.keys })
-            expect(copySpy).to.have.been.calledOnce
         })
     })
